+++ run_demo.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env sh
 set -eu
-python -m triagebloom analyze sample_data/demo_events.json --output-dir reports
+python -m triagebloom analyze sample_data/combined_incident.json --output-dir reports --disable-off-hours --redact